Correction supp Utilisateur

+++ src/UtilisateurBD.java
@@ -46,14 +46,16 @@ public class UtilisateurBD {
 
     public void supprimerUtilisateur(int num) throws SQLException {
         ResultSet resultNumObj = laConnexionMySQL.createStatement().executeQuery("SELECT idob FROM OBJET WHERE idut =" + num + ";");
-        resultNumObj.next();
-        int numObj = resultNumObj.getInt(1);
-        ResultSet resultNumVe = laConnexionMySQL.createStatement().executeQuery("SELECT idve FROM VENTE WHERE idob =" + numObj + ";");
-        resultNumVe.next();
-        int numVe = resultNumVe.getInt(1);
-        laConnexionMySQL.createStatement().executeUpdate("DELETE FROM ENCHERIR WHERE idve =" + numVe + ";");
-        laConnexionMySQL.createStatement().executeUpdate("DELETE FROM VENTE WHERE idob =" + numObj + ";");
-        laConnexionMySQL.createStatement().executeUpdate("DELETE FROM PHOTO WHERE idob =" + numObj + ";");
+        if (resultNumObj.next()) {
+            int numObj = resultNumObj.getInt(1);
+            ResultSet resultNumVe = laConnexionMySQL.createStatement().executeQuery("SELECT idve FROM VENTE WHERE idob =" + numObj + ";");
+            if (resultNumVe.next()) {
+                int numVe = resultNumVe.getInt(1);
+                laConnexionMySQL.createStatement().executeUpdate("DELETE FROM ENCHERIR WHERE idve =" + numVe + ";");
+            }
+            laConnexionMySQL.createStatement().executeUpdate("DELETE FROM VENTE WHERE idob =" + numObj + ";");
+            laConnexionMySQL.createStatement().executeUpdate("DELETE FROM PHOTO WHERE idob =" + numObj + ";");
+        }
         laConnexionMySQL.createStatement().executeUpdate("DELETE FROM OBJET WHERE idut =" + num + ";");
         laConnexionMySQL.createStatement().executeUpdate("DELETE FROM ENCHERIR WHERE idut =" + num + ";");
         laConnexionMySQL.createStatement().executeUpdate("DELETE FROM UTILISATEUR WHERE idut =" + num + ";");
